pyproject.toml
[project]
name = "mcprobe"
version = "0.1.0"
description = "A testing framework that validates MCP servers provide sufficient information for LLM agents to answer real-world questions correctly"
readme = "README.md"
license = {text = "GPL-3.0"}
requires-python = ">=3.11"
authors = [
    {name = "MCProbe Contributors"}
]
keywords = ["mcp", "testing", "llm", "agents", "model-context-protocol"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.0",
    "httpx>=0.27",
    "ollama>=0.4",
]

[project.optional-dependencies]
adk = [
    "google-adk>=1.22.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "respx>=0.21",
    "pytest-mock>=3.12",
    "mypy>=1.10",
    "ruff>=0.6",
    "types-PyYAML>=6.0",
]

[project.scripts]
mcprobe = "mcprobe.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcprobe"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "e2e: end-to-end tests requiring external services (Ollama)",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W",      # pycodestyle
    "F",           # pyflakes
    "I",           # isort
    "N",           # pep8-naming
    "UP",          # pyupgrade
    "B",           # bugbear
    "C4",          # comprehensions
    "SIM",         # simplify
    "ARG",         # unused arguments
    "PTH",         # pathlib
    "ERA",         # commented-out code
    "PL",          # pylint
    "RUF",         # ruff-specific
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["mcprobe"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["PLR2004"]  # Allow magic values in tests for clarity

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "ollama.*",
    "google.adk.*",
    "google.genai.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "mcprobe.agents.adk"
disallow_untyped_calls = false

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = ["method-assign"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
